pyproject.toml
[project]
name = "api-slug"
version = "0.1.0"
description = "Create slug from url"
authors = [
    {name = "AnKov00",email = "ankov00@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[uvicorn] (>=0.122.0,<0.123.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "python-jose[cryptogrphy] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
